# Optimize timestamp queries: ==, !=, !NULL