--- a/src/realm/query_engine.cpp
+++ b/src/realm/query_engine.cpp
@@ -821,11 +821,97 @@ size_t TimestampNode<LessEqual>::find_first_local(size_t start, size_t end)
 
     return not_found;
 }
+
+template <>
+size_t TimestampNode<Equal>::find_first_local(size_t start, size_t end)
+{
+    REALM_ASSERT(this->m_table);
+
+    if (m_value.is_null()) {
+        if (REALM_UNLIKELY(!m_condition_column_is_nullable)) {
+            return not_found;
+        }
+        return this->find_first_local_seconds<Equal>(start, end);
+    }
+
+    while (start < end) {
+        size_t ret = this->find_first_local_seconds<Equal>(start, end);
+
+        if (ret == not_found)
+            return not_found;
+
+        // We now know that neither m_value nor current value is null and that seconds part equals
+        // We are just missing to compare nanoseconds part
+        int32_t nanos = this->get_nanoseconds_and_cache(ret);
+        if (nanos == m_value.get_nanoseconds()) {
+            return ret;
+        }
+        start = ret + 1;
+    }
+
+    return not_found;
+}
+
+template <>
+size_t TimestampNode<NotEqual>::find_first_local(size_t start, size_t end)
+{
+    REALM_ASSERT(this->m_table);
+
+    // in many scenarios it is likely that the first item is not equal do a quick first check
+    if (start < end) {
+        util::Optional<int64_t> seconds = get_seconds_and_cache(start);
+        if (seconds != m_needle_seconds
+            || (seconds && this->get_nanoseconds_and_cache(start) != m_value.get_nanoseconds())) {
+            return start;
+        }
+    }
+
+    ++start;
+
+    if (m_value.is_null()) {
+        if (REALM_UNLIKELY(!m_condition_column_is_nullable)) {
+            return not_found;
+        }
+        return this->find_first_local_seconds<NotNull>(start, end);
+    }
+
+    int64_t needle_seconds = m_value.get_seconds();
+    while (start < end) {
+        util::Optional<int64_t> seconds = get_seconds_and_cache(start);
+        if (!seconds || *seconds != needle_seconds) {
+            return start;
+        }
+        // We now know that neither m_value nor current value is null and that seconds part equals
+        // We are just missing to compare nanoseconds part
+        int32_t nanos = this->get_nanoseconds_and_cache(start);
+        if (nanos != m_value.get_nanoseconds()) {
+            return start;
+        }
+        ++start;
+    }
+
+    return not_found;
+}
+
+template <>
+size_t TimestampNode<NotNull>::find_first_local(size_t start, size_t end)
+{
+    REALM_ASSERT(this->m_table);
+    if (REALM_UNLIKELY(!m_condition_column_is_nullable)) {
+        return start; // all are not null, return first
+    }
+    return this->find_first_local_seconds<NotNull>(start, end);
+}
+
+
 #ifdef _WIN32
 // Explicit instantiation required on some windows builds
 template size_t TimestampNode<Greater>::find_first_local(size_t start, size_t end);
 template size_t TimestampNode<Less>::find_first_local(size_t start, size_t end);
 template size_t TimestampNode<GreaterEqual>::find_first_local(size_t start, size_t end);
 template size_t TimestampNode<LessEqual>::find_first_local(size_t start, size_t end);
+template size_t TimestampNode<Equal>::find_first_local(size_t start, size_t end);
+template size_t TimestampNode<NotEqual>::find_first_local(size_t start, size_t end);
+template size_t TimestampNode<NotNull>::find_first_local(size_t start, size_t end);
 #endif
 } // namespace realm