@jedelbo 2019-08-16
I was able to squeeze it a bit more. A simple optimization in `Array::find_optimized` paid well off.
These are the improvements compared to the commit that gave the numbers above.
```
QueryTimestampGreater (MemOnly, EncryptionOff):     min   7.80ms (-5.85%)            max   7.93ms (-4.78%)            med   7.81ms (-6.06%)            avg   7.82ms (-5.89%)            stddev    36us (+114.87%)     
QueryTimestampGreater (MemOnly, EncryptionOn):      min   7.82ms (-5.51%)            max   7.96ms (-4.28%)            med   7.91ms (-4.66%)            avg   7.91ms (-4.62%)            stddev    37us (+224.19%)     
QueryTimestampGreater (Full   , EncryptionOff):     min   7.81ms (-5.74%)            max   7.85ms (-7.33%)            med   7.83ms (-6.54%)            avg   7.83ms (-6.71%)            stddev    11us (-86.54%)      
QueryTimestampGreater (Full   , EncryptionOn):      min   7.82ms (-5.70%)            max   7.89ms (-5.45%)            med   7.84ms (-5.65%)            avg   7.84ms (-5.61%)            stddev    18us (+13.93%)      

QueryTimestampGreaterEqual (MemOnly, EncryptionOff):     min   7.64ms (-8.30%)            max   7.73ms (-8.67%)            med   7.64ms (-8.33%)            avg   7.65ms (-8.34%)            stddev    26us (-32.48%)      
QueryTimestampGreaterEqual (MemOnly, EncryptionOn):      min   7.64ms (-8.19%)            max   7.75ms (-7.53%)            med   7.66ms (-8.06%)            avg   7.68ms (-7.94%)            stddev    40us (+94.54%)      
QueryTimestampGreaterEqual (Full   , EncryptionOff):     min   7.62ms (-9.13%)            max   7.69ms (-11.20%)           med   7.63ms (-10.79%)           avg   7.64ms (-10.22%)           stddev    19us (-82.88%)      
QueryTimestampGreaterEqual (Full   , EncryptionOn):      min   7.64ms (-8.00%)            max   7.69ms (-8.91%)            med   7.66ms (-8.61%)            avg   7.66ms (-8.47%)            stddev    14us (-66.94%)      

QueryTimestampLess (MemOnly, EncryptionOff):             min   7.66ms (-9.38%)            max   7.72ms (-8.82%)            med   7.69ms (-9.16%)            avg   7.68ms (-9.18%)            stddev    20us (+241.73%)     
QueryTimestampLess (MemOnly, EncryptionOn):              min   7.70ms (-9.25%)            max   7.73ms (-9.73%)            med   7.72ms (-9.70%)            avg   7.71ms (-9.56%)            stddev    10us (-58.71%)      
QueryTimestampLess (Full   , EncryptionOff):             min   7.68ms (-9.22%)            max   7.76ms (-8.68%)            med   7.71ms (-9.02%)            avg   7.70ms (-9.08%)            stddev    27us (+53.37%)      
QueryTimestampLess (Full   , EncryptionOn):              min   7.68ms (-9.26%)            max   7.75ms (-8.65%)            med   7.70ms (-9.12%)            avg   7.70ms (-9.08%)            stddev    26us (+168.02%)     

QueryTimestampLessEqual (MemOnly, EncryptionOff):        min   7.73ms (-6.85%)            max   7.78ms (-6.50%)            med   7.75ms (-6.63%)            avg   7.75ms (-6.68%)            stddev    19us (+120.61%)     
QueryTimestampLessEqual (MemOnly, EncryptionOn):         min   7.74ms (-6.98%)            max   7.78ms (-7.23%)            med   7.76ms (-6.98%)            avg   7.76ms (-7.01%)            stddev    16us (-17.55%)      
QueryTimestampLessEqual (Full   , EncryptionOff):        min   7.73ms (-6.92%)            max   7.81ms (-7.99%)            med   7.75ms (-8.06%)            avg   7.75ms (-7.82%)            stddev    23us (-66.96%)      
QueryTimestampLessEqual (Full   , EncryptionOn):         min   7.75ms (-7.05%)            max   7.84ms (-6.63%)            med   7.79ms (-6.75%)            avg   7.79ms (-6.79%)            stddev    33us (+75.85%)      

QueryTimestampEqual (MemOnly, EncryptionOff):            min    882us (-40.98%)           max    938us (-38.51%)           med    884us (-40.94%)           avg    885us (-40.93%)           stddev     6us (+0.78%)       
QueryTimestampEqual (MemOnly, EncryptionOn):             min    893us (-40.74%)           max    900us (-43.82%)           med    895us (-40.79%)           avg    895us (-41.29%)           stddev     2us (-93.73%)      
QueryTimestampEqual (Full   , EncryptionOff):            min    882us (-40.99%)           max    967us (-39.88%)           med    884us (-41.10%)           avg    887us (-41.56%)           stddev    11us (-63.36%)      
QueryTimestampEqual (Full   , EncryptionOn):             min    894us (-40.69%)           max    934us (-42.52%)           med    895us (-42.43%)           avg    897us (-42.03%)           stddev     5us (-82.14%)      

QueryTimestampNotEqual (MemOnly, EncryptionOff):         min  13.43ms (-54.78%)           max  13.59ms (-54.49%)           med  13.54ms (-54.55%)           avg  13.50ms (-54.67%)           stddev    59us (+19.57%)      
QueryTimestampNotEqual (MemOnly, EncryptionOn):          min  13.44ms (-54.89%)           max  13.58ms (-54.56%)           med  13.54ms (-54.61%)           avg  13.50ms (-54.73%)           stddev    56us (+73.53%)      
QueryTimestampNotEqual (Full   , EncryptionOff):         min  13.39ms (-55.16%)           max  13.54ms (-55.09%)           med  13.52ms (-54.81%)           avg  13.47ms (-55.01%)           stddev    64us (-25.45%)      
QueryTimestampNotEqual (Full   , EncryptionOn):          min  13.47ms (-55.60%)           max  13.64ms (-56.20%)           med  13.60ms (-56.09%)           avg  13.56ms (-56.07%)           stddev    68us (-70.62%)      

QueryTimestampNotNull (MemOnly, EncryptionOff):          min  14.25ms (-5.77%)            max  14.31ms (-8.73%)            med  14.30ms (-5.77%)            avg  14.28ms (-6.17%)            stddev    28us (-83.88%)      
QueryTimestampNotNull (MemOnly, EncryptionOn):           min  14.19ms (-5.96%)            max  14.28ms (-7.43%)            med  14.24ms (-6.14%)            avg  14.23ms (-6.28%)            stddev    31us (-75.08%)      
QueryTimestampNotNull (Full   , EncryptionOff):          min  14.23ms (-5.67%)            max  14.30ms (-6.65%)            med  14.26ms (-5.80%)            avg  14.26ms (-5.86%)            stddev    26us (-65.96%)      
QueryTimestampNotNull (Full   , EncryptionOn):           min  14.20ms (-6.02%)            max  14.29ms (-5.86%)            med  14.27ms (-5.82%)            avg  14.26ms (-5.88%)            stddev    26us (+14.03%)      

QueryTimestampEqualNull (MemOnly, EncryptionOff):        min   3.87ms (-9.52%)            max   3.90ms (-9.52%)            med   3.88ms (-9.34%)            avg   3.88ms (-9.38%)            stddev     8us (-6.84%)       
QueryTimestampEqualNull (MemOnly, EncryptionOn):         min   3.87ms (-9.61%)            max   3.90ms (-9.13%)            med   3.89ms (-9.51%)            avg   3.89ms (-9.51%)            stddev     7us (+139.19%)     
QueryTimestampEqualNull (Full   , EncryptionOff):        min   3.87ms (-9.65%)            max   3.90ms (-9.42%)            med   3.88ms (-9.31%)            avg   3.88ms (-9.40%)            stddev    10us (+90.95%)      
QueryTimestampEqualNull (Full   , EncryptionOn):         min   3.88ms (-9.66%)            max   3.90ms (-10.66%)           med   3.89ms (-9.46%)            avg   3.89ms (-9.60%)            stddev     7us (-59.88%)      
```